+++ services/ui-src/src/labels/QualifierFormsDatas.tsx
@@ -1,0 +1,4 @@
+export * as QualifierFormsData2021 from "./2021/qualifierFormsData";
+export * as QualifierFormsData2022 from "./2022/qualifierFormsData";
+export * as QualifierFormsData2023 from "./2023/qualifierFormsData";
+export * as QualifierFormsData2024 from "./2024/qualifierFormsData";
