+++ services/ui-src/src/measures/2022/shared/Qualifiers/index.tsx
@@ -5,7 +5,7 @@ import * as CUI from "@chakra-ui/react";
 import { Data } from "labels/2022/qualifierFormsData";
 import { useEffect, useState } from "react";
 import { validationFunctions } from "shared/Qualifiers/validationFunctions";
-import { DeliverySystems } from "./deliverySystems";
+import { DeliverySystems } from "shared/Qualifiers/deliverySystems";
 import { useParams } from "react-router-dom";
 import * as Types from "types";
